test(bootstrap): refactor buildProposal for re-use
